Move attribute out of its own folder
