+++ lib/attribute.js
@@ -1,8 +1,8 @@
 // Highly modeled after the interface for YUI's Attribute.
 // Though much simpler internally. 
 
-var ObjectUtils = require('./../utilities/objects');
-var Lang = require('./../utilities/lang');
+var ObjectUtils = require('./utilities/objects');
+var Lang = require('./utilities/lang');
 
 var OPTS = {
   GETTER: 'getter',
