+++ lib/utilities/createClass.js
@@ -1,5 +1,5 @@
 var ObjectUtils = require('./objects');
-var Attribute = require('./../attribute/attribute');
+var Attribute = require('./../attribute');
 var augment = require('./augment');
 
 var defaultConfig = {
